cheri: Fix compile in drivers/comedi
- Pass a uintptr_t to the comedi_timeout callback instead of
  and unsigned long. Add a coccinelle script to do the fixup.
- Other unsigned long vs. uintptr_t issues.
- Standard ioctl issues.

+++ drivers/comedi/drivers.c
@@ -298,8 +298,8 @@ int comedi_timeout(struct comedi_device *dev,
 		   int (*cb)(struct comedi_device *dev,
 			     struct comedi_subdevice *s,
 			     struct comedi_insn *insn,
-			     unsigned long context),
-		   unsigned long context)
+			     uintptr_t context),
+		   uintptr_t context)
 {
 	unsigned long timeout = jiffies + msecs_to_jiffies(COMEDI_TIMEOUT_MS);
 	int ret;
